backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# Database — plain psycopg2, no ORM
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation
pydantic==2.7.1
pydantic-settings==2.3.1
email-validator==2.1.1
